Time: 699 ms (14.73%), Space: 99.5 MB (6.27%) - LeetHub

+++ 1726-tuple-with-same-product/1726-tuple-with-same-product.cpp
@@ -1,6 +1,6 @@
 class Solution {
-    int fact(int x){
-       int sum=1;
+    long long fact(int x){
+       long long sum=1;
         for(int i=1;i<=x;i++){
            sum*=i;
         }
